+++ templates/content.php
@@ -6,7 +6,4 @@
   <div class="entry-summary">
     <?php the_excerpt(); ?>
   </div>
-  <footer>
-    <?php the_tags('<ul class="entry-tags"><li>','</li><li>','</li></ul>'); ?>
-  </footer>
 </article>